+++ src/task/mod.rs
@@ -49,6 +49,7 @@ pub trait Task: Send + Sync + std::marker::Sized {
     /// Default task options.
     const DEFAULTS: TaskOptions = TaskOptions {
         time_limit: None,
+        hard_time_limit: None,
         max_retries: None,
         min_retry_delay: None,
         max_retry_delay: None,